\name{info.Map-class}
\docType{class}
\alias{info.Map}
\alias{info.Map-class}
\alias{initialize,info.Map-method}

\title{Class for information about Map/Operators}
\description{
Auxiliar class for definition of an object of class \code{\link{yuima.Map}}. see the documentation of \code{\link{yuima.Map}} for more details.
}

