\name{info.Ppr}
\docType{class}
\alias{info.Ppr}
\alias{info.Ppr-class}
\alias{initialize,info.Ppr-method}


\title{Class for information about Point Process}
\description{
Auxiliar class for definition of an object of class \code{\link{yuima.Ppr}} and \code{\link{yuima.Hawkes}}. see the documentation for more details.
}
