% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{initialize,yuima.ae-method}
\alias{initialize,yuima.ae-method}
\alias{initialize,yuima.ae,ANY-method}
\title{Constructor for \code{yuima.ae} Class}
\usage{
\S4method{initialize}{yuima.ae}(
  .Object,
  order,
  var,
  u.var,
  eps.var,
  characteristic,
  Z0,
  Mu,
  Sigma,
  c.gamma,
  verbose
)
}
\arguments{
\item{.Object}{an object of class \code{\link{yuima.ae-class}}.}

\item{order}{integer. The order of the expansion.}

\item{var}{character. The state variables.}

\item{u.var}{character. The variables of the characteristic function.}

\item{eps.var}{character. The perturbation variable.}

\item{characteristic}{expression. The characteristic function.}

\item{Z0}{numeric. The solution to the deterministic process obtained by setting the perturbation to zero.}

\item{Mu}{numeric. The drift vector for the representation of Z1.}

\item{Sigma}{matrix. The diffusion matrix for the representation of Z1.}

\item{c.gamma}{list. The coefficients of the Hermite polynomials.}

\item{verbose}{logical. Print on progress? Default \code{FALSE}.}
}
\description{
Construct an object of class \code{\link{yuima.ae-class}}.
}
