\name{Integral.sde}
\docType{class}
\alias{Integral.sde}
\alias{Integral.sde-class}
%\alias{yuima.Integral}
%\alias{info.Output}
%\alias{param.Output}
\alias{initialize,Integral.sde-method}

\title{Class for the mathematical description of integral of a stochastic process}

\description{
Auxiliar class for definition of an object of class \code{\link{yuima.Integral}}. see the documentation of \code{\link{yuima.Integral}} for more details.
}

