\name{Intensity.PPR}
\alias{Intensity.PPR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Intesity Process for the Point Process Regression Model}
\description{This function returns the intensity process of a Point Process Regression Model}
\usage{
Intensity.PPR(yuimaPPR, param)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{yuimaPPR}{An object of class \code{yuima.PPR}}
  \item{param}{Model parameters}
}
\value{On obejct of class \code{yuima.data}}
\author{YUIMA TEAM}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
#INSERT HERE AN EXAMPLE
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
