\name{lambdaFromData}
\alias{lambdaFromData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Intensity of a Point Process Regression Model
}
\description{This function returns the intensity process of a Ppr model when covariates and counting processes are obsered on discrete time}
\usage{
lambdaFromData(yuimaPpr, PprData = NULL, parLambda = list())
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{yuimaPpr}{ Mathematical Description of Ppr model
}
  \item{PprData}{Observed data}
  \item{parLambda}{Values of intesity parameters}
}
\details{ ... }
\value{...}
\references{
...}
\author{YUIMA TEAM}
\note{...}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
...}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
