\name{mmfrac}
\alias{mmfrac}
\title{
mmfrac
}
\description{
Estimates the drift of s fractional Ornstein-Uhlenbeck and, if necessary, also the Hurst and diffusion parameters.
}
\usage{
mmfrac(yuima, ...)
}
\arguments{
  \item{yuima}{a \code{yuima} object.
}
  \item{\dots}{arguments passed to \code{\link{qgv}}.
}
}
\details{
Estimates the drift of s fractional Ornstein-Uhlenbeck and, if necessary, also the Hurst and diffusion parameters.
}
\value{an object of class mmfrac
}
\references{
Brouste, A., Iacus, S.M. (2013) Parameter estimation for the discretely observed fractional Ornstein-Uhlenbeck process and the Yuima R package, Computational Statistics, pp. 1129--1147.
}
\author{The YUIMA Project Team}
\seealso{
See also \code{\link{qgv}}.
}
\examples{
#Example
}
% R documentation directory.
\keyword{ts}
