\name{param.Integral}
\docType{class}
\alias{param.Integral}
\alias{param.Integral-class}
%\alias{yuima.Integral}
%\alias{info.Output}
%\alias{param.Output}
\alias{initialize,param.Integral-method}

\title{Class for the Mathematical Description of Integral of a Stochastic Process}

\description{
Auxiliar class for definition of an object of class \code{\link{yuima.Integral}}. see the documentation of \code{\link{yuima.Integral}} for more details.
}

