\name{param.Output-class}
\docType{class}
\alias{param.Output}
\alias{param.Output-class}
\alias{initialize,param.Output-method}

\title{Class for information about Map/Operators}
\description{
Auxiliar class for definition of an object of class \code{\link{yuima.Output}}. see the documentation of \code{\link{yuima.Output}} for more details.
}

