% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{plot,yuima.ae,ANY-method}
\alias{plot,yuima.ae,ANY-method}
\alias{plot,yuima.ae-method}
\title{Plot Method for \code{yuima.ae} Class}
\usage{
\S4method{plot}{yuima.ae,ANY}(x, grids = list(), eps = 1, order = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{yuima.ae-class}}.}

\item{grids}{list. A named list of vectors specifying the grid to evaluate the density. The names must match the state variables.}

\item{eps}{numeric. The intensity of the perturbation.}

\item{order}{integer. The expansion order. If \code{NULL} (default), it uses the maximum order used in \code{ae}.}

\item{...}{additional arguments passed to the plot function.}
}
\description{
Plot an object of class \code{\link{yuima.ae-class}}.
}
