\name{setHawkes}
\alias{setHawkes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Constructor of Hawkes model}
\description{
...
}
\usage{
setHawkes(lower.var = "0", upper.var = "t", var.dt = "s",
  process = "N", dimension = 1, intensity = "lambda",
  ExpKernParm1 = "c", ExpKernParm2 = "a", const = "nu",
  measure = NULL, measure.type = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lower.var}{...}
  \item{upper.var}{...}
  \item{var.dt}{...}
  \item{process}{...}
  \item{dimension}{...}
  \item{intensity}{...}
  \item{ExpKernParm1}{...}
  \item{ExpKernParm2}{...}
  \item{const}{...}
  \item{measure}{...}
  \item{measure.type}{...}
}
\details{...}
\value{...
}
\references{
...}
\author{
...}
\note{
...}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{...}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
