\name{setPPR}
\alias{setPPR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Point Process}
\description{Constructor of a Point Process }
\usage{
setPPR(yuima, counting.var = "N", gFun, Kernel,
  var.dx = "s", var.dt = "s", lambda.var = "lambda",
  lower.var = "0", upper.var = "t", nrow = 1, ncol = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{yuima}{...}
  \item{counting.var}{...}
  \item{gFun}{...}
  \item{Kernel}{...}
  \item{var.dx}{...}
  \item{var.dt}{...}
  \item{lambda.var}{...}
  \item{lower.var}{...}
  \item{upper.var}{...}
  \item{nrow}{...}
  \item{ncol}{...}
}
\details{...}
\value{...}
\references{...}
\author{...}
\note{...}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
...}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
