\name{subsampling}
\alias{subsampling}
\title{subsampling }
\description{subsampling }
\usage{
subsampling(x,y)
}
\arguments{
  \item{x}{an \code{\link{yuima-class}} or 
    \code{\link{yuima.model-class}} object.}
  \item{y}{initial value vector of state variables.}
 }
\value{
  \item{yuima}{a \code{yuima-class} object.}
}
\author{The YUIMA Project Team}
\note{This is curently an empty function. To be implemented soon.
}
\keyword{ts}
