\name{ybook}
\title{R code for the Yuima Book}
\alias{ybook}
\description{
  Shows the R code corresponding to each chapter in the Yuima Book.
}
\usage{
ybook(chapter)
}
\arguments{
\item{chapter}{a number in 1:7}
}
\details{
  This is an accessory function which open the R code corresponding to
  Chapter "chapter" in the Yuima Book so that the reader can replicate
  the code.
}
\examples{
ybook(1)
}
\keyword{misc}
