\name{yuima.characteristic-class}
\docType{class}
\alias{yuima.characteristic-class}
\alias{initialize,yuima.characteristic-method}
\title{Classe for stochastic differential equations characteristic scheme}
\description{
  The \code{yuima.characteristic} class is a class of the \pkg{yuima} package.  
}
\section{Slots}{
  \describe{
    \item{\code{equation.number}:}{The number of equations modeled in
	  the \code{yuima} object.}
    \item{\code{time.scale}:}{The time scale assumed in the \code{yuima}
	  object.}
  }
}
\author{The YUIMA Project Team}
\keyword{classes}
