\name{yuima.functional-class}
\docType{class}
\alias{yuima.functional-class}

\alias{getF,yuima.functional-method}
\alias{getf,yuima.functional-method}
\alias{getxinit,yuima.functional-method}
\alias{gete,yuima.functional-method}
\alias{initialize,yuima.functional-method}


\title{Classes for stochastic differential equations data object}
\description{
  The \code{yuima.functional} class is a class of the  \pkg{yuima} package.  
}


\author{YUIMA Project}
\keyword{classes}
