\name{yuima.Hawkes}
\docType{class}
\alias{yuima.Hawkes}
\alias{yuima.Hawkes-class}
\alias{initialize,yuima.Hawkes-method}
\alias{simulate,yuima.Hawkes-method}

\title{Class for a mathematical description of a Point Process}
\description{
The \code{yuima.Hawkes-class} is a class of the \pkg{yuima} package that extends the \code{\link{yuima.PPR-class}}. The object of this class contains all the information about the Hawkes process with exponential kernel. 

An object of this class can be created by calls of the function \code{\link{setHawkes}}.
}
