\name{yuima.law-class}
\docType{class}
\alias{yuima.law-class}
\alias{yuima.law}
\alias{initialize,yuima.law-method}


\title{Class of yuima law}

\description{Insert Description Here}

\section{Slots}{...:
\describe{
  \item{rng}{
  function}
    \item{density}{
  function}
    \item{cdf}{
  function}
    \item{param.measure}{...}
    \item{characteristic}{
  function
  }
    \item{time.var}{
  label}
    \item{rand.var}{
  label}
    \item{charact.var}{
  label}
    \item{dim}{
  number}
}
}

\section{Methods}{...}

\author{The YUIMA Project Team}
\keyword{classes}

