\name{yuima.model-class}
\docType{class}
\alias{yuima.model-class}
\alias{initialize,yuima.model-method}
\alias{initialize,model.parameter-method}
\alias{limiting.gamma,yuima.model-method}
%%\alias{setSampling,yuima.model-method}

\title{Classes for the mathematical description of stochastic differential equations}
\description{
  The \code{yuima.model} class is a class of the  \pkg{yuima} package.  
}
\section{Slots}{
  \describe{
    \item{\code{drift}:}{add explanation}
    \item{\code{diffusion}:}{add explanation}
    \item{\code{jump.coeff}:}{add explanation}
    \item{\code{measure}:}{add explanation}
    \item{\code{measure.type}:}{add explanation}
    \item{\code{parameter}:}{add explanation}
    \item{\code{state.variable}:}{add explanation}
    \item{\code{jump.variable}:}{add explanation}
    \item{\code{time.variable}:}{add explanation}
    \item{\code{noise.number}:}{add explanation}
    \item{\code{equation.number}:}{add explanation}
    \item{\code{dimension}:}{add explanation}
    \item{\code{solve.variable}:}{add explanation}
    \item{\code{xinit}:}{add explanation}
  }
}
\author{YUIMA Project}
\keyword{classes}
