\name{info.PPR}
\docType{class}
\alias{info.PPR}
\alias{info.PPR-class}
\alias{initialize,info.PPR-method}


\title{Class for information about Point Process}
\description{
Auxiliar class for definition of an object of class \code{\link{yuima.PPR}} and \code{\link{yuima.Hawkes}}. see the documentation for more details.
}
