\encoding{UTF-8}
\name{LawMethods}
\alias{LawMethods}
\alias{rand}
\alias{rand-method}
\alias{cdf}
\alias{quant}
\alias{char}
\alias{dens}

\title{Methods for an object of class \code{yuima.law}}
\description{Methods for yuima.law}
\usage{
rand(object, n, param, ...)
dens(object, x, param, log = FALSE, ...)
cdf(object, q, param, ...)
quant(object, p, param, ...)
}
\arguments{
  \item{object}{...}
  \item{n}{...}
  \item{param}{...}
  \item{...}{...}
  \item{x}{...}
  \item{log}{...}
  \item{q}{...}
  \item{p}{...}
}

\value{
Methods for an object of yuima.law-class}
\author{
YUIMA TEAM}
\note{
Insert additional info}
