\name{yuima.sampling-class}
\docType{class}
\alias{yuima.sampling-class}
\alias{initialize,yuima.sampling-method}
\title{Classes for stochastic differential equations sampling scheme}
\description{
  The \code{yuima.sampling} class is a class of the \pkg{yuima} package.  
}
\section{Slots}{
  \describe{
    \item{\code{Terminal}:}{Horizon of the stochastic processes.}
    \item{\code{division}:}{Time division for the data and/or model.}
  }
}
\author{YUIMA Project}
\keyword{classes}
